build on 32 bit linux. addresses #1195
git-svn-id: file:///svn/toku/tokudb.1195@7644 c7de825b-a66e-492c-adef-691d508d4ae1

+++ newbrt/tests/cachetable-unpin-and-remove-test.c
@@ -51,7 +51,7 @@ cachetable_unpin_and_remove_test (int n) {
     // put the keys into the cachetable
     for (i=0; i<n; i++) {
         u_int32_t hi = toku_cachetable_hash(f1, make_blocknum(keys[i].b));
-        r = toku_cachetable_put(f1, make_blocknum(keys[i].b), hi, (void *) keys[i].b, 1, flush, fetch, 0);
+        r = toku_cachetable_put(f1, make_blocknum(keys[i].b), hi, (void *)(long) keys[i].b, 1, flush, fetch, 0);
         assert(r == 0);
     }
